moved loggers from middleware to utils
