+++ src/__tests__/config/mongoMemoryServer.ts
@@ -1,5 +1,5 @@
 import mongoose from 'mongoose'
-import logger from '../../middlewares/winstonLogger'
+import logger from '../../utils/winstonLogger'
 import { MongoMemoryServer } from 'mongodb-memory-server'
 
 let mongoMemoryServer: MongoMemoryServer
